requirements.txt
aiodns==3.0.0
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.1.1
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
blinker==1.7.0
blosc2==2.0.0
boto3==1.26.83
botocore==1.29.165
bs4==0.0.1
ccxt==2.4.66
celery==5.3.1
certifi==2023.11.17
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.4
configobj==5.0.8
configparser==6.0.1
constantly==15.1.0
contourpy==1.1.0
croniter==1.4.1
cryptography==38.0.4
cycler==0.11.0
daphne==4.0.0
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash-table==5.0.0
DateTime==5.4
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.52.0
django-cors-headers==3.10.1
django-drf-filepond==0.4.1
django-filter==21.1
django-storages==1.13.2
djangorestframework==3.14.0
etelemetry==0.3.1
execnb==0.1.5
executing==1.2.0
fastcore==1.5.29
fastjsonschema==2.18.0
filelock==3.8.0
finvizfinance==0.14.7
fitz==0.0.1.dev2
flags==0.0.1.2
Flask==3.0.0
fonttools==4.42.0
fqdn==1.5.1
frozendict==2.3.10
frozenlist==1.3.3
gevent==23.7.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.8.0
incremental==22.10.0
instaloader==4.10
intelhex==2.3.0
investpy==1.0.8
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.24
jsonpointer==2.4
jsonschema==4.16.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.25.4
kiwisolver==1.4.4
kombu==5.3.1
looseversion==1.3.0
lxml==5.0.0
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mercury==2.3.1
mistune==3.0.1
mplfinance==0.12.10b0
msgpack==1.0.4
multidict==6.0.4
multitasking==0.0.11
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.7
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.2
oauthlib==3.2.2
overrides==7.7.0
packaging==22.0
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
peewee==3.17.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2022.8.24
platformdirs==2.5.2
plotly==5.16.1
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==4.21.12
prov==2.0.0
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pydot==2.0.0
pyee==8.2.2
Pygments==2.13.0
PyJWT==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyppeteer==1.0.2
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3.post1
pywatchman==1.4.1
pyxnat==1.6.2
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.5.1
QtPy==2.4.1
rdflib==7.0.0
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
s3transfer==0.6.2
schedule==1.2.1
scipy==1.12.0
Send2Trash==1.8.2
service-identity==23.1.0
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.36
sqlparse==0.4.4
stack-data==0.6.2
swagger-spec-validator==2.7.6
TA-Lib==0.4.28
tenacity==8.2.3
terminado==0.18.1
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
traits==6.3.2
tvdatafeed @ git+https://github.com/rongardF/tvdatafeed.git@e6f6aaa7de439ac6e454d9b26d2760ded8dc4923
Twisted==22.10.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.7.1
tzdata==2023.4
Unidecode==1.3.8
uri-template==1.3.0
urllib3==1.26.16
vine==5.0.0
virtualenv==20.16.3
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
Werkzeug==3.0.1
whitenoise==6.5.0
widgetsnbextension==4.0.8
yarl==1.8.2
yfinance==0.2.33
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
